Added multithreading for apply_search
The method apply_search now uses multiple threads to filter and color
entries faster. The max amount of threads is defined by the constant
MAX_THREADS_AMOUNT and is currently 20.

From now on commit titles will be capitalized. Sorry for the change of
style in the middle of the project :(